APIMF-2823: ensure parsing of double does not lose precision

+++ shared/src/main/scala/amf/plugins/document/graph/parser/ExpandedGraphParser.scala
@@ -296,7 +296,7 @@ class ExpandedGraphParser()(implicit val ctx: GraphParserContext) extends GraphP
         case Type.Int =>
           instance.setWithoutId(f, int(node), annotations(nodes, sources, key))
         case Type.Float =>
-          instance.setWithoutId(f, float(node), annotations(nodes, sources, key))
+          instance.setWithoutId(f, double(node), annotations(nodes, sources, key))
         case Type.Double =>
           instance.setWithoutId(f, double(node), annotations(nodes, sources, key))
         case Type.DateTime =>
